requirements.txt
appdirs==1.4.3
certifi==2017.4.17
click==6.7
future==0.16.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-dotenv==0.6.4
python-telegram-bot==5.3.1
six==1.10.0
SQLAlchemy==1.1.9
urllib3==1.20
